# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Gym CRM API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/members/{member_id}/ai-plans": {
      "get": {
        "tags": [
          "ai-plans"
        ],
        "summary": "List Plans",
        "operationId": "list_plans_api_members__member_id__ai_plans_get",
        "parameters": [
          {
            "name": "member_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Member Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AIPlanOut"
                  },
                  "title": "Response List Plans Api Members  Member Id  Ai Plans Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ai-plans"
        ],
        "summary": "Generate",
        "description": "Starts generating a new plan version. Poll the plan until it's no longer 'generating'.",
        "operationId": "generate_api_members__member_id__ai_plans_post",
        "parameters": [
          {
            "name": "member_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Member Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateIn"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIPlanOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ai-plans/{plan_id}": {
      "get": {
        "tags": [
          "ai-plans"
        ],
        "summary": "Get Plan",
        "operationId": "get_plan_api_ai_plans__plan_id__get",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIPlanOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "ai-plans"
        ],
        "summary": "Update Plan",
        "description": "Trainer edits. Published plans can be edited too; members see changes right away.",
        "operationId": "update_plan_api_ai_plans__plan_id__patch",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Plan Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AIPlanUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIPlanOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ai-plans"
        ],
        "summary": "Delete Plan",
        "operationId": "delete_plan_api_ai_plans__plan_id__delete",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Plan Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ai-plans/{plan_id}/publish": {
      "post": {
        "tags": [
          "ai-plans"
        ],
        "summary": "Publish Plan",
        "operationId": "publish_plan_api_ai_plans__plan_id__publish_post",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIPlanOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ai-plans/{plan_id}/unpublish": {
      "post": {
        "tags": [
          "ai-plans"
        ],
        "summary": "Unpublish Plan",
        "operationId": "unpublish_plan_api_ai_plans__plan_id__unpublish_post",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIPlanOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AIPlanOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "member_id": {
            "type": "string",
            "format": "uuid",
            "title": "Member Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "status": {
            "$ref": "#/components/schemas/PlanStatus"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params"
          },
          "content": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlanContent"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "created_by": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StaffRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "member_id",
          "version",
          "status",
          "title",
          "params",
          "content",
          "error",
          "model",
          "created_by",
          "published_at",
          "created_at",
          "updated_at"
        ],
        "title": "AIPlanOut"
      },
      "AIPlanUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "content": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlanContent"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "AIPlanUpdate"
      },
...
      "DietPlan": {
        "properties": {
          "daily_calories": {
            "type": "integer",
            "title": "Daily Calories"
          },
          "protein_g": {
            "type": "integer",
            "title": "Protein G"
          },
          "carbs_g": {
            "type": "integer",
            "title": "Carbs G"
          },
          "fat_g": {
            "type": "integer",
            "title": "Fat G"
          },
          "meals": {
            "items": {
              "$ref": "#/components/schemas/Meal"
            },
            "type": "array",
            "title": "Meals"
          },
          "hydration": {
            "type": "string",
            "title": "Hydration"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "daily_calories",
          "protein_g",
          "carbs_g",
          "fat_g",
          "meals",
          "hydration",
          "notes"
        ],
        "title": "DietPlan"
      },
...
      "Exercise": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "sets": {
            "type": "integer",
            "title": "Sets"
          },
          "reps": {
            "type": "string",
            "title": "Reps"
          },
          "rest_seconds": {
            "type": "integer",
            "title": "Rest Seconds"
          },
          "notes": {
            "type": "string",
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "name",
          "sets",
          "reps",
          "rest_seconds",
          "notes"
        ],
        "title": "Exercise"
      },
...
      "GenerateIn": {
        "properties": {
          "days_per_week": {
            "type": "integer",
            "maximum": 6.0,
            "minimum": 2.0,
            "title": "Days Per Week",
            "default": 4
          },
          "session_minutes": {
            "type": "integer",
            "maximum": 120.0,
            "minimum": 30.0,
            "title": "Session Minutes",
            "default": 60
          },
          "equipment": {
            "type": "string",
            "enum": [
              "full_gym",
              "basic",
              "bodyweight"
            ],
            "title": "Equipment",
            "default": "full_gym"
          },
          "instructions": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Instructions"
          }
        },
        "type": "object",
        "title": "GenerateIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Meal": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "time": {
            "type": "string",
            "title": "Time"
          },
          "items": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Items"
          },
          "calories": {
            "type": "integer",
            "title": "Calories"
          },
          "protein_g": {
            "type": "integer",
            "title": "Protein G"
          }
        },
        "type": "object",
        "required": [
          "name",
          "time",
          "items",
          "calories",
          "protein_g"
        ],
        "title": "Meal"
      },
...
      "PlanContent": {
        "properties": {
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "weekly_schedule": {
            "items": {
              "$ref": "#/components/schemas/WorkoutDay"
            },
            "type": "array",
            "title": "Weekly Schedule"
          },
          "diet": {
            "$ref": "#/components/schemas/DietPlan"
          },
          "progression": {
            "type": "string",
            "title": "Progression"
          },
          "safety_notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Safety Notes"
          }
        },
        "type": "object",
        "required": [
          "summary",
          "weekly_schedule",
          "diet",
          "progression",
          "safety_notes"
        ],
        "title": "PlanContent"
      },
...
      "PlanStatus": {
        "type": "string",
        "enum": [
          "generating",
          "failed",
          "draft",
          "published",
          "archived"
        ],
        "title": "PlanStatus"
      },
...
      "StaffRef": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "StaffRef"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WorkoutDay": {
        "properties": {
          "day": {
            "type": "string",
            "title": "Day"
          },
          "focus": {
            "type": "string",
            "title": "Focus"
          },
          "warmup": {
            "type": "string",
            "title": "Warmup"
          },
          "exercises": {
            "items": {
              "$ref": "#/components/schemas/Exercise"
            },
            "type": "array",
            "title": "Exercises"
          },
          "cooldown": {
            "type": "string",
            "title": "Cooldown"
          }
        },
        "type": "object",
        "required": [
          "day",
          "focus",
          "warmup",
          "exercises",
          "cooldown"
        ],
        "title": "WorkoutDay"
      }
    }
  }
}